<refactor> 클래스의 역할을 반영한 패키지 구조 변경
