+++ src/main/java/webserver/http/message/HttpResponse.java
@@ -1,7 +1,5 @@
 package webserver.http.message;
 
-import webserver.utils.HttpHeaderUtils;
-
 import java.util.Arrays;
 import java.util.Objects;
