+++ src/main/java/webserver/http/message/parser/body/RequestBodyParser.java
@@ -1,4 +1,4 @@
-package webserver.utils.parser.body;
+package webserver.http.message.parser.body;
 
 import java.util.Map;
