+++ src/main/java/webserver/http/message/parser/body/RequestBodyParserManager.java
@@ -1,4 +1,4 @@
-package webserver.utils.parser.body;
+package webserver.http.message.parser.body;
 
 import webserver.exceptions.BadRequestException;
